+++ src/lib/index.ts
@@ -1,4 +1,4 @@
-import { jurinetLib } from './jurinetLib';
+import { jurinetLib } from './jurinet';
 import { juricaLib } from './juricaLib';
 
 export { jurinetLib, juricaLib };
